Added more Authentication

+++ flask-api/resources/item.py
@@ -1,6 +1,7 @@
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort
 from sqlalchemy.exc import SQLAlchemyError
+from flask_jwt_extended import jwt_required, get_jwt
 
 from db import db
 from schemas import ItemSchema, ItemUpdateSchema
@@ -11,13 +12,20 @@ blp = Blueprint("items", __name__, description="Operations on Items")
 
 @blp.route("/item/<int:item_id>")
 class Item(MethodView):
+    @jwt_required()
     @blp.response(200, ItemSchema)
     def get(self, item_id):
         # use primary key
         item = ItemModel.query.get_or_404(item_id)
         return item
 
+    @jwt_required()
     def delete(self, item_id):
+        jwt = get_jwt()
+
+        if not jwt.get("is_admin"):
+            abort(401, "Admin privilege is required.")
+
         item = ItemModel.query.get_or_404(item_id)
 
         db.session.delete(item)
@@ -47,6 +55,7 @@ class ItemList(MethodView):
     def get(self):
         return ItemModel.query.all()
 
+    @jwt_required(fresh=True)
     @blp.arguments(ItemSchema)
     @blp.response(201, ItemSchema)
     def post(self, item_data):
